+++ src/main/java/lotto/service/LottoIssueService.java
@@ -3,6 +3,7 @@ package lotto.service;
 import lotto.domain.LottoMachine;
 import lotto.domain.Lottos;
 import lotto.domain.Money;
+import lotto.service.dto.LottosDTO;
 
 public class LottoIssueService {
     private final LottoMachine lottoMachine = new LottoMachine();
